FEATURE : created a command to update pokemon images url in db
